Enemy Spawner Thingy
I made a makeshift boss thing in my script. The DroneRespawnerScript (not the best name but it is what it is) handles most of the logic. It looks very similar to DroneSpawnerScript, but it is  tailored specifically to what is needed for the task. The BossMovementTempScript is a very rudimentary script and was solely for testing, so it is not necessary to the actual objective of the task.